requirements.txt
transformers
torch
gradio
newspaper3k
lxml_html_clean
